change bat field to float and corresponding migration

+++ maps/models.py
@@ -21,7 +21,7 @@ class Pollution(models.Model):
     co2 = models.IntegerField()
     smoke = models.FloatField()
     noise = models.IntegerField()
-    bat = models.IntegerField()
+    bat = models.FloatField()
     received = models.DateTimeField()
     device = models.ForeignKey(Device, null=True)
